requirement.txt
Pillow              9.2.0
djangorestframework 3.14.0
Django              4.1.2